fix: selector filter error
