+++ console/src/domain/dataset/components/DatasetLabel.tsx
@@ -22,6 +22,17 @@ export const DatasetLabelText = themedStyled('div', () => ({
     textOverflow: 'ellipsis',
 }))
 
+export function getDatastLabel(version: IDatasetTreeVersionSchema, dataset?: IDatasetTreeSchema) {
+    const p = dataset ? [dataset.ownerName, dataset.projectName, dataset.datasetName].join('/') : ''
+    const name = version?.versionName ?? version?.name
+    const v = (name ?? '').substring(0, 8)
+    const title = [p, v, version.alias, version.createdTime ? formatTimestampDateTime(version.createdTime) : '']
+        .filter((tmp) => !!tmp)
+        .join('/')
+
+    return title
+}
+
 export function DatasetLabel({
     version,
     dataset,
@@ -38,7 +49,7 @@ export function DatasetLabel({
     const p = dataset ? [dataset.ownerName, dataset.projectName, dataset.datasetName].join('/') : ''
     const name = version?.versionName ?? version?.name
     const v = (name ?? '').substring(0, 8)
-    const title = [p, v, version.createdTime ? formatTimestampDateTime(version.createdTime) : '']
+    const title = [p, v, version.alias, version.createdTime ? formatTimestampDateTime(version.createdTime) : '']
         .filter((tmp) => !!tmp)
         .join('/')
